% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapped_CI.R
\name{bootstrapped_CI}
\alias{bootstrapped_CI}
\title{Produce bootstrapped 95% Confidence Interval}
\usage{
bootstrapped_CI(data_vector, n_resamples = 1000)
}
\arguments{
\item{data_vector}{observed data to sample from}

\item{n_resamples}{number of resamples to take}
}
\value{
tibble
}
\description{
Produce bootstrapped 95% Confidence Interval
}
