% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_libraries.R
\name{load_libraries}
\alias{load_libraries}
\title{Load all packages used in the book}
\usage{
load_libraries()
}
\description{
Load all packages used in the book
}
\details{
Currently, these are:
\itemize{
\item{tidyverse}
\item{brms}
\item{ggsignif}
\item{gridExtra}
\item{Rlab}
\item{rcartocolor}
\item{extraDistr}
\item{boot}
\item{naniar}
}
}
