% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ida.R
\name{theme_ida}
\alias{theme_ida}
\title{ggplot theme}
\usage{
theme_ida(title.size = 16, text.size = 14, legend.position = "top",
  show.axis = FALSE, plot.margin = c(0.2, 0.1, 0.2, 0.1))
}
\arguments{
\item{title.size}{font size of all title elements + facet labels}

\item{text.size}{font size of all text elements}

\item{legend.position}{either one of "top", "bottom", "left" or "right"}

\item{show.axis}{should axes be drawn? (TRUE, FALSE, "x", "y")}

\item{plot.margin}{margin around entire plot (in cm of the top, right, bottom, and left margins)}
}
\description{
ggplot theme
}
